Added "Save to PDF" button
Added a "Save to PDF" button that makes use of the "window.print()" built in JavaScript function. This function however does not select print as PDF by default. The user must manually select that option. In the future I would like to make printing to a pdf the default option